+++ CHANGELOG.md
@@ -7,9 +7,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Added
 - Now the subscription returned by the `subscription()` method in the `PlanSubscriber` trait is resolved by a class compatible with `Gerardojbaez\Laraplans\Contracts\SubscriptionResolverInterface`. Default resolver is `Gerardojbaez\Laraplans\SubscriptionResolver`. Behavior and logic not changed. [Documentation](https://laraplans.readthedocs.io/en/latest/usage.html#subscription-resolving).
+- `Gerardojbaez\Laraplans\Events\SubscriptionSaved`
+- `Gerardojbaez\Laraplans\Events\SubscriptionSaving`
 
 ### Changed
 - Now using `Event::dispatch()` instead of `Event::fire()`. PR #49. Requires **Laravel 5.8 or newer.**
+
+### Removed
+- Method `Gerardojbaez\Laraplans\Models\PlanSubscription::boot()`, logic was moved to event listeners.
 
 ## [2.2.0] - 2018-02-23